Add files via upload
Decided to upload via the UI on the website. For some reason when I uploaded from VSCode I wasn't getting a checkbox, and, truth be told I like the dopamine hit I get when I see the boxes that I did the work! This was the conclusion of the Flexbox section in the Bootcamp.